% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{get_stac_type}
\alias{get_stac_type}
\title{Get the STAC type of a Value}
\usage{
get_stac_type(value)
}
\arguments{
\item{value}{The Value to get the type of}
}
\value{
An external pointer
}
\description{
Get the STAC type of a Value
}
