% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{rust_get_type_name}
\alias{rust_get_type_name}
\title{Get the STAC type of a Value}
\usage{
rust_get_type_name(value)
}
\arguments{
\item{value}{The Value to get the type of}
}
\value{
An external pointer
}
\description{
Get the STAC type of a Value
}
