% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stacio-package.R
\docType{package}
\name{stacio-package}
\alias{stacio}
\alias{stacio-package}
\title{stacio: Bindings to the 'STAC' Rust Crate for Reading and Writing SpatioTemporal Asset Catalog Objects}
\description{
What the package does (one paragraph).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mikemahoney218.github.io/stacio/}
}

}
\author{
\strong{Maintainer}: Michael Mahoney \email{mike.mahoney.218@gmail.com} (\href{https://orcid.org/0000-0003-2402-304X}{ORCID})

}
\keyword{internal}
